third_party/blink/web_tests/platform/generic/storage/indexeddb/createIndex-after-failure-expected.txt

Test IndexedDB's basics.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "createIndex-after-failure.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
objectStore.createIndex('index', 'key', {unique: true})
objectStore.deleteIndex('index')
Expecting exception from objectStore.deleteIndex('index')
PASS Exception was thrown.
PASS code is DOMException.NOT_FOUND_ERR
PASS ename is 'NotFoundError'
Exception message: Failed to execute 'deleteIndex' on 'IDBObjectStore': The specified index was not found.
Now requesting object2
now we wait.
deleteIndexAfterGetError()
Expecting exception from objectStore.deleteIndex('index')
PASS Exception was thrown.
PASS code is 11
PASS ename is 'InvalidStateError'
Exception message: Failed to execute 'deleteIndex' on 'IDBObjectStore': The object store has been deleted.
Expecting exception from objectStore.deleteIndex('index')
PASS Exception was thrown.
PASS code is 11
PASS ename is 'InvalidStateError'
Exception message: Failed to execute 'deleteIndex' on 'IDBObjectStore': The object store has been deleted.
PASS successfullyParsed is true

TEST COMPLETE
